Refactor worker image rendering and move image coverage into worker-specific tests
